backend/pyproject.toml
[project]
name = "app-evaluation-agent"
version = "1.0.0"
description = "AI-based automated application evaluation tool."
authors = [
    {name = "lumingye",email = "lumingye@zj.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.120.1,<0.121.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "toml (>=0.10.2,<0.11.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "arq (>=0.26.3,<0.27.0)",
    "alembic (>=1.17.0,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "python-dotenv (>=1.0.0,<2.0.0)",
    "openai (>=2.6.1,<3.0.0)",
    "pillow (>=10.0.0,<11.0.0)",
    "websockets (>=15.0.1,<16.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poe.tasks]
migrate = "alembic upgrade head"
worker = "arq app_evaluation_agent.worker.WorkerSettings"


[tool.poe.tasks.start]
cmd = "uvicorn app_evaluation_agent.main:app --reload --port 9010 --host 0.0.0.0"
env = { HF_ENDPOINT = "https://hf-mirror.com" }

[tool.poe.tasks.start_debug]
cmd = "uvicorn app_evaluation_agent.main:app --reload --port 9010 --host 0.0.0.0 --log-level debug"
env = { HF_ENDPOINT = "https://hf-mirror.com", APP_LOG_LEVEL = "DEBUG" }


[[tool.poetry.source]]
name = "tuna_mirror"
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"
priority = "primary"


[dependency-groups]
dev = [
    "poethepoet (>=0.37.0,<0.38.0)",
    "pytest (>=7.4.0,<9.0.0)",
    "pytest-asyncio (>=0.23.0,<0.24.0)",
    "aiosqlite (>=0.20.0,<0.21.0)",
    "ruff (>=0.6.0,<0.7.0)",
    "black (>=24.4.0,<25.0.0)"
]
